Start migration of FluentCommandLineBuilder<T> to FluentCommandLineParser<T>
The Builder should be renamed to the same as the FluentCommandLineParser so that they are seen as the same command line parser. It would be nice to have just the one name. I have made the Builder obsolete without error and brought in a generic version of the FluentCommandLineParser. When the consumer news up a FCLP they should see there is the standard version and a generic version - perhaps this will increase adoption of the generic variant.

The FluentCommandLineBuilder will go through a couple of stages

1. Made obsolete with no error
2. Made obsolete with error.
3. Delete

The timescales for this are unknown at this stage.